src/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
defusedxml==0.7.1
Django==3.2
django-appconf==1.0.5
django-ckeditor==6.7.0
django-compressor==4.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-cuser==2017.3.16
django-environ==0.11.2
django-hosts==5.2
django-js-asset==2.1.0
django-taggit==4.0.0
django-tinymce==3.6.1
django-tqdm==1.3.1
django-translations==1.3.0
django-user-agents==0.4.0
djangorestframework==3.14.0
fontawesomefree==6.4.2
idna==3.4
modeltranslation==0.25
Pillow==10.0.0
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
stripe==6.4.0
tqdm==4.66.1
typing_extensions==4.7.1
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
